Updated events and services to use ptrees

+++ util/src/util/platform/linux/file_linux.c
@@ -12,7 +12,7 @@ file_load_into_memory(const char* file_name, void** buffer, file_opts_e opts)
     int fd;
     struct stat stbuf;
     uintptr_t buffer_size;
-    
+
     /* get file pointer */
     if(opts & FILE_BINARY)
         fp = fopen(file_name, "rb");
@@ -23,7 +23,7 @@ file_load_into_memory(const char* file_name, void** buffer, file_opts_e opts)
         fprintf(stderr, "fopen() failed for file \"%s\"\n", file_name);
         return 0;
     }
-    
+
     /* get file descriptor and stat file */
     fd = open(file_name, O_RDONLY);
     if(fd == -1)
@@ -36,17 +36,17 @@ file_load_into_memory(const char* file_name, void** buffer, file_opts_e opts)
         fprintf(stderr, "fstat() failed for file \"%s\"\n", file_name);
         return 0;
     }
-    
+
     /* ensure file is regular file, as stbuf.st_size is only valid for regular files */
     if(!S_ISREG(stbuf.st_mode))
     {
         fprintf(stderr, "File \"%s\" is not a regular file\n", file_name);
         return 0;
     }
-    
+
     /* compute file size */
     buffer_size = (uintptr_t)stbuf.st_size;
-    
+
     /* allocate buffer to copy into */
     if(opts & FILE_BINARY)
         *buffer = MALLOC(buffer_size);
@@ -57,9 +57,9 @@ file_load_into_memory(const char* file_name, void** buffer, file_opts_e opts)
         fprintf(stderr, "malloc() failed in file_load_into_memory() -- not enough memory");
         return 0;
     }
-    
+
     /* copy file into buffer */
-    if((intptr_t)fread(*buffer, 1, buffer_size, fp) != buffer_size)
+    if(fread(*buffer, 1, buffer_size, fp) != buffer_size)
     {
         fprintf(stderr, "fread() failed for file \"%s\"\n", file_name);
         return 0;
@@ -69,7 +69,7 @@ file_load_into_memory(const char* file_name, void** buffer, file_opts_e opts)
     /* append null terminator if not in binary mode */
     if((opts & FILE_BINARY) == 0)
         ((char*)(*buffer))[buffer_size] = '\0';
-    
+
     return (uint32_t)buffer_size;
 }
